permute the check on manifest
Previously we checked if the manifest received was the same as the old
one first, and if not we checked if the new manifest is what we want.

This had the unfortunate consequence of not allowing updates to the same
target the device is already at (and forever blocking updates). So these
checks were permuted.

I tried to clean up the code as well, this code is still a bit tricky
though.