[PRO-3454] publish UpdateSpec event
In order to update the status of the device in the device registry
we need to publish an `UpdateSpec` event for the multi target update
updates. Normally this is done by core, but for multi target updates
core is not involved so we(the director) need to do it.

Notice that the UpdateSpec event contains a field that is called
`packageUuid` which doesn't make any sense for the multi target update,
this field is not used neither by the device registry nor the ui. So I'm
filling it with 00000000-0000-0000-0000-000000000000, which of
course feels wrong, but it will work for now. The `UpdateSpec`
event will probably be removed in the future, so abusing the
`packageUuid` field is probably okay...

Also notice that the codecs for `Instant` is different between rvi-sota
and libats, so I created a SOTA_Instant type that will use the old
codecs.