[PRO-3361] change format for device manifest
This change is made to be more compliant with the uptane spec, we still
accept the old format for backward compatability reasons.

Also note that we at the device manifest level don't require that what
is in the ecu-manifest actually is an ecu-manifest. This is checked
later in the verifcation of the device-manifest. This is so that an
error in one of the ecu-manifests doesn't break the whole device
manifest.