[PRO-3581] targets.json version should increase
When something goes wrong the director sets the new target whawt the
device had before the errorneous update. The way this happened was
that a new timestamp.json was created that pointed to the older
target.json version. But this is not correct, we need to create a new
version of the target.json, with the new version number.