[PRO-3472] check signatures on the received json
Previously we parsed the json and then checked, but this causes problems
with checking the signature. So instead the endpoint takes an
SignedPayload[Json], and we check the signature against the Json rather
than the DeviceManifest, which we manually parse from the Json during
verification of the device manifest.

Also I changed the type of the hashes from ClientHashes to my own Hashes
type, that also only accept SHA256 at the moment.

I still don't really know how we want to implement
multiple hashes in the future, but the current Map
forces two things:

1. The server needs to know all hashmethods that a
  client can possibly send (with all what that means
  for backwards compatibility)

2. All hashes need to have the same length as a SHA256
  hash. Because ValidChecksum checks the length, this
  is of course easy to change.