Increase akka connection pool size
During two P2 incedents we noticed that director was using a very low
cpu ammount even though each pod was assigned with a large cpu limit.

We noticed that slick db queues were increasing but db connections
were still being handled and not dropped.

We noticed that the akka pool was set to only 8 connections maximum,
which is not enough to handle many connections and explains the low
cpu usage even though there is a high cpu limit.

This commit increases the akka pool limits to handle more connections
at the same time.

the number of threads for the db is also doubled as to reduce the
slick queue size, even though we don't that is the limiting factor.

Signed-off-by: Simão Mata <simao.mata@here.com>