Always bump version when refreshing metadata
In some cases the metadata version is not bumped when refreshing a role expiration date.

For example:

1. device is at `current_target` 5
2. device downloads targets version 6 with new
update (`current_target` remains at 5)
3. metadata expires before device installs it
4. device requests targets.json again, but is provided version 6 again
with a different expire date. This happens because the version is
calculated based on the version the device has
installed (`current_target`), rather than the version the device actually downloaded last (`current_target + 1`).

This commit calculates makes director calculate the version bump based on the last generated metadata (`file_cache` table), rather than the version that was used to install the last update (`device_current_targets` table)

Additionally, when the device finally pushes a manifest, the version is
bumped to the last generated targets.json, rather than
`current_target`. This is not strictly necessary since the version
used to generate metadata would now be in `file_cache`, but it would
mean the versions of `current_target` and `file_cache` would never match again.

Signed-off-by: Simão Mata <simao.mata@here.com>