Force role regeneration if roles were updated
Due to the fact that sometimes roles are updated (if we had done this
properly this would never happen, but we are using insertOrUpdate when
persisting roles, which means we can generate the same version of a
role with different signed contents), it can be the case that some
devices receive the same metadata version with different contents.

This change forces the regeneration of a bumped version of the
metadata when `updated_at` is bigger than `created_at` by more than 5
seconds. This is not optimal but is the only way I can see to detect
that the metadata was updated. Using simply `updated_at > created_at`
is not enough because our API allows a user to create two
different file cache requests at roughly the same time, when they are
processed the last one wins.

Signed-off-by: Simão Mata <simao.mata@here.com>