Take latest generated role into account for device assignments
Take latest generated role into account every time current assignment
is calculated.

This is a follow up to https://github.com/advancedtelematic/director/pull/198

Without this change, targets are renewed and version is bumped, but
the targets could be empty in some cases.

This is another consequence of using the role version to determine
which software the device is running, this means we need to always
check which version is higher, the one used for assignments or the one
used to generate roles, and use the assignment version to generate a
new role version.

Signed-off-by: Simão Mata <simao.mata@here.com>