Force targets regeneration if devices got wrong targets
This forces targets.json regeneration if previous targets should have
had correlationId but did not.

This is for devices that downloaded a regenerated targets before
d3d4acbc was merged.

Additionaly, also add correlation id if device did not have a
targets.json in cache, this is due to bugfix/OTA-4539.

Signed-off-by: Simão Mata <simao.mata@here.com>