Move signed roles migration to a main method
This is so it can use more than 1 connection to speed up the
processes.

This, along with batching, does make it faster, but it still takes
several hours, possibly days.

For this reason, implemented support for NULL checksum and length in
signed roles, which means the migration is not need, and can be
executed later if needed.

For this to work, roles fetched from the db that need to be converted
to a signed role, will trigger a canonical json calculations. This
only happens when fetching the last db role of a device, so it will
happen once for each device as they connect to the backend, subsequent
calls by the device will not trigger this calculation.

Signed-off-by: Simão Mata <simao.mata@here.com>