Allow assignment to be created when device has serial of another device
In some cases we were not scoping the queries to `ecus` using
`deviceId`, just `ecu_serial`, but the PK for that table is
`(device_id,ecu_serial)`, which means those queries would return
results for other devices. One of the consequences was that an
assignment could not be created if there was an assignment for another
device that had the same ecu serial.

Signed-off-by: Simão Mata <simao.mata@here.com>