Send DeviceUpdateCompleted even if device did not send install report
Some legacy devices are not sending an install report if some
secondary ecu sends a report, so we need to look deeper into the
manifest to check if some ecu reported some installation.

Since those install reports also do not contain any correlation id, we
need to infer this correlation id from the processed assignments and
so we need to send one message for each processed assignment.

Signed-off-by: Simão Mata <simao.mata@here.com>