Replace primary ECUs deterministically.
Since there is only one primary ECU per device, if we replace it we can know
which is the new ECU that is the replacement. That is not the case for the
secondaries: if we replace more than one secondary, the system has no way of
knowing which ECU is taking the place of which.

So, when publishing the ECU replacement events, for primaries we can determine
which is the replacement and which is the old ECU. For secondaries, we can only
arbitrarily match some of the old secondaries with some of the new one.

Relates-to: OTA-5130
Signed-off-by: Ignacio Lucero <ignacio.lucero@here.com>